+++ pycmplot/__init__.py
@@ -42,4 +42,4 @@ __all__ = [
     "ResourceConfig",
 ]
 
-__version__ = "0.1.8"
+__version__ = "0.1.9"